+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/uday/Desktop/kaya/tic-tac-toe/src/game.js":"1","/Users/uday/Desktop/kaya/tic-tac-toe/src/square.js":"2","/Users/uday/Desktop/kaya/tic-tac-toe/src/board.js":"3"},{"size":271,"mtime":1609395915334,"results":"4","hashOfConfig":"5"},{"size":176,"mtime":1609395714527,"results":"6","hashOfConfig":"5"},{"size":1525,"mtime":1609395823026,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ah0pgf",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/uday/Desktop/kaya/tic-tac-toe/src/game.js",[],"/Users/uday/Desktop/kaya/tic-tac-toe/src/square.js",[],"/Users/uday/Desktop/kaya/tic-tac-toe/src/board.js",[]]
+[{"/Users/uday/Desktop/kaya/tic-tac-toe/src/game.js":"1","/Users/uday/Desktop/kaya/tic-tac-toe/src/square.js":"2","/Users/uday/Desktop/kaya/tic-tac-toe/src/board.js":"3"},{"size":1257,"mtime":1609399761371,"results":"4","hashOfConfig":"5"},{"size":183,"mtime":1609399447730,"results":"6","hashOfConfig":"5"},{"size":611,"mtime":1609399566882,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ah0pgf",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/uday/Desktop/kaya/tic-tac-toe/src/game.js",[],"/Users/uday/Desktop/kaya/tic-tac-toe/src/square.js",[],"/Users/uday/Desktop/kaya/tic-tac-toe/src/board.js",[]]
